feat(AB#37384): fixes
