+++ packages/angular/package.json
@@ -22,12 +22,12 @@
     "@angular/core": "^21.1.0"
   },
   "dependencies": {
+    "lodash": "^4.17.23",
     "tslib": "^2.3.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~21.1.0",
     "@types/lodash": "^4.17.24",
-    "lodash": "^4.17.23",
     "@angular/cli": "~21.1.0",
     "@angular/common": "^21.1.0",
     "@angular/compiler": "^21.1.0",
