+++ packages/angular/src/atoms/Stack.ts
@@ -25,43 +25,17 @@ export type Props = {
 };
 
 import { tv } from 'tailwind-variants';
-import {
-  Direction,
-  AlignItems,
-  Justify,
-  getAlignItems,
-  getJustify,
-  resolveResponsive,
-  resolveResponsiveVariants,
-} from './utilities';
-import type { ResponsiveValue } from './utilities';
-export const stackVariants = tv({
+import { Direction, AlignItems, Justify } from './constants';
+import type { ResponsiveValue } from './constants';
+import { getAlignItems, getJustify, resolveResponsive } from './utilities';
+const directionToClass = (direction: string, bp?: string): string => {
+  const cls = direction === 'row' ? 'gi-flex-row' : 'gi-flex-col';
+  return bp ? `${bp}:${cls}` : cls;
+};
+const gapToClass = (gap: number, bp?: string): string => (bp ? `${bp}:gi-gap-${gap}` : `gi-gap-${gap}`);
+const stackVariants = tv({
   base: ['gi-flex', 'gi-w-full'],
   variants: {
-    direction: {
-      row: 'gi-flex-row',
-      column: 'gi-flex-col',
-    },
-    smDirection: {
-      row: 'sm:gi-flex-row',
-      column: 'sm:gi-flex-col',
-    },
-    mdDirection: {
-      row: 'md:gi-flex-row',
-      column: 'md:gi-flex-col',
-    },
-    lgDirection: {
-      row: 'lg:gi-flex-row',
-      column: 'lg:gi-flex-col',
-    },
-    xlDirection: {
-      row: 'xl:gi-flex-row',
-      column: 'xl:gi-flex-col',
-    },
-    '2xlDirection': {
-      row: '2xl:gi-flex-row',
-      column: '2xl:gi-flex-col',
-    },
     align: {
       start: 'gi-items-start',
       center: 'gi-items-center',
@@ -83,7 +57,6 @@ export const stackVariants = tv({
     },
   },
   defaultVariants: {
-    direction: 'column',
     align: 'start',
     justify: 'start',
     wrap: false,
@@ -101,11 +74,14 @@ export const stackVariants = tv({
       [ngStyle]="styles"
       [class]="
         stackVariants({
-          ...resolveResponsiveVariants(direction ?? Direction.COLUMN, 'Direction'),
           align: getAlignItems(align),
           justify: getJustify(justify),
           wrap: wrap ?? false,
-          class: [resolveResponsive(gap ?? 0, 'gi-gap'), className],
+          class: [
+            resolveResponsive(direction ?? Direction.COLUMN, directionToClass),
+            resolveResponsive(gap ?? 0, gapToClass),
+            className,
+          ],
         })
       "
       [attr.data-testid]="dataTestId"
@@ -124,21 +100,22 @@ export const stackVariants = tv({
   imports: [CommonModule],
 })
 export default class Stack {
+  directionToClass = directionToClass;
+  gapToClass = gapToClass;
   stackVariants = stackVariants;
   getAlignItems = getAlignItems;
   getJustify = getJustify;
   resolveResponsive = resolveResponsive;
-  resolveResponsiveVariants = resolveResponsiveVariants;
 
   @Input() id!: Props['id'];
   @Input() role!: Props['role'];
   @Input() ariaLabel!: Props['ariaLabel'];
   @Input() ariaLabelledBy!: Props['ariaLabelledBy'];
   @Input() styles!: Props['styles'];
-  @Input() direction!: Props['direction'];
   @Input() align!: Props['align'];
   @Input() justify!: Props['justify'];
   @Input() wrap!: Props['wrap'];
+  @Input() direction!: Props['direction'];
   @Input() gap!: Props['gap'];
   @Input() className!: Props['className'];
   @Input() dataTestId!: Props['dataTestId'];
