+++ packages/angular/tailwind.config.ts
@@ -6,7 +6,11 @@ const config: Config = {
   content: ['./src/**/*.{ts,html}', './.storybook/**/*.{ts,html}'],
   theme: createTheme(),
   plugins: [require('@tailwindcss/typography')],
-  safelist: [{ pattern: /gap-./, variants: ['xs', 'sm', 'md', 'lg', 'xl', '2xl'] }],
+  safelist: [
+    { pattern: /gap-./, variants: ['xs', 'sm', 'md', 'lg', 'xl', '2xl'] },
+    { pattern: /flex-row/, variants: ['xs', 'sm', 'md', 'lg', 'xl', '2xl'] },
+    { pattern: /flex-col/, variants: ['xs', 'sm', 'md', 'lg', 'xl', '2xl'] },
+  ],
 };
 
 export default config;
