+++ packages/core/atoms/Button.lite.tsx
@@ -1,6 +1,6 @@
 import { tv } from 'tailwind-variants';
 import { useMetadata } from '@builder.io/mitosis';
-import { Variant, Appearance, Size } from './utilities';
+import { Variant, Appearance, Size } from './constants';
 
 export const ButtonSize = {
   SM: Size.SM,
