+++ packages/react/src/stories/Button.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import _ from 'lodash';
 import { Button } from '../atoms/index.js';
 import { ButtonSize } from '../atoms/Button';
-import { Variant, Appearance } from '../atoms/utilities.js';
+import { Variant, Appearance } from '../atoms/constants.js';
 import {
   buttonMeta,
   Default as buttonDefault,
